Fix pointer cast warnings on 32-bit ESP32
Pin amy submodule to merged PR #571 (2339c88) which uses intptr_t
instead of int64_t for void* state casts in patches.c.

Also fix tulip_helpers get_stack_pointer to use uintptr_t instead of
uint64_t for the same reason.

+++ tulip/shared/tulip_helpers.c
@@ -25,9 +25,9 @@ int check_rx_char() {
 }
 #endif
 
-uint64_t get_stack_pointer() {
+uintptr_t get_stack_pointer() {
     char dummy[64];
-    uint64_t val = (uint64_t)dummy;
+    uintptr_t val = (uintptr_t)dummy;
     return val;
 }
