+++ tulip/shared/tulip_helpers.h
@@ -26,5 +26,5 @@ uint32_t tulip_fread(mp_obj_t file, uint8_t * buf, uint32_t len);
 void tulip_fclose(mp_obj_t file);
 uint32_t tulip_fseek(mp_obj_t file, uint32_t seekpoint);
 int32_t tulip_getline(char * line, uint32_t * len, mp_obj_t file );
-uint64_t get_stack_pointer();
+uintptr_t get_stack_pointer();
 #endif
